backend endpoints, controlers  and routes refactor
