+++ backend/routes/journey.routes.js
@@ -3,7 +3,6 @@ const journeyRouter = express.Router();
 import { userAuth } from "../middlewares/users.middleware.js";
 import { 
   getInTransit, 
-  newContainer, 
   createJourney, 
   createJourneyLog,
   getJourneyLog,
@@ -14,7 +13,6 @@ import {
   updateJourney
 } from "../controllers/journey.controller.js";
 
-journeyRouter.post('/container', userAuth, newContainer)
 journeyRouter.post('/journey', userAuth, createJourney);
 journeyRouter.patch('/journey', updateJourney);
 journeyRouter.get('/in-transit', userAuth, getInTransit);
